refactor: BaseActivity
