+++ app/src/main/java/com/sdhong/pokemonapp/base/BaseActivity.kt
@@ -1,0 +1,30 @@
+package com.sdhong.pokemonapp.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseActivity<VB : ViewBinding>(
+    private val bindingFactory: (layoutInflater: LayoutInflater) -> VB
+) : AppCompatActivity() {
+
+    protected lateinit var binding: VB
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        binding = bindingFactory(layoutInflater)
+        setContentView(binding.root)
+
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+    }
+}
