+++ app/src/main/java/com/sdhong/pokemonapp/view/DetailActivity.kt
@@ -3,34 +3,22 @@ package com.sdhong.pokemonapp.view
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import coil3.load
 import com.sdhong.pokemonapp.R
+import com.sdhong.pokemonapp.base.BaseActivity
 import com.sdhong.pokemonapp.databinding.ActivityDetailBinding
 import com.sdhong.pokemonapp.util.collectLatestStateFlow
 import com.sdhong.pokemonapp.viewmodel.DetailViewModel
 
-class DetailActivity : AppCompatActivity() {
+class DetailActivity : BaseActivity<ActivityDetailBinding>(
+    bindingFactory = ActivityDetailBinding::inflate
+) {
 
-    private lateinit var binding: ActivityDetailBinding
     private val viewModel: DetailViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        binding = ActivityDetailBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
 
         setCollectors()
     }
